Add a testtype for running Skia tests.
The test is not necessarily Skia specific, but it provides
functionality that allows native Skia tests to be run.

Includes options to specify the Skia test app to run (inside
nativetest directory), flags to pass to the test app, and a file
to retrieve off the device after the test completes. (Skia test
apps record their results to a json file, so retrieving this file
allows us to view the results so long as the app completed.)

Delay running the test long enough to start capturing logcat.
Without logcat, we cannot know if the app crashed.

Send the results json file to the ITestInvocationListener for
viewing. Report success or failure based on whether the file
existed (since failing to create the file means we likely crashed,
and at the very least do not have results to look at).

BUG=skbug.com/2454

Change-Id: Ica21dd2068bd2d299db1dab0d3c7d13e9a0dc53b