Change the default value of "require-battery-check" option to true.
Because of this option is set to false default, many of tests in the lab are
being allocated to devices with low battery even though they have "min-battery"
option specified in the command lines. Considering that normally providing
"min-battery" or "max-battery" option mean these requirements to be applied in all cases,
it is reasonable to have "require-battery-check" to be true by default.

Bug: 18428691
Change-Id: I1aa81bec28af5126045ceee628b5d94722e47743