Capture the emulator stdout and stderr, update the log.
Add a new runCmdInBackground with SizeLimitedOutputStream to the RunUtil,
so the output of the command will be captured in the
SizeLimitedOutputStream.
Add getEmulatorOutput to the ITestDevice,
so the emulator command output can be fetched with the TestDevice.
Add setEmulatorOutputStream to the TestDevice, so later the emulator
output can be fetched.
Add stopEmulatorOutputStream to stop stream emulator output.
Add a new option verbose to SdkAvdPreparer, this will enable emulator
verbose output.

Change-Id: I0e2c08ef663452d39f78547397dbee509944bb8b