Solving unittests failures.
In ag/663471, we add stopEmulatorOutputStream to TestDevice which will
be used in DeviceManger.killEmulator(). But in unittest we use
IManagedTestDevice to mock emulator, some tests fail due to the mocked
device can't be cast to TestDevice.
Some tests fail due to a NullPointerException in
TestInvocation.reportLogs(), as we use device.getIDevice().isEmulator()
here. But in unittest, the mock device return null for
device.getIDevice(), which is not the case for real device or emulator.

Moving stopEmulatorOutputStream to ITestDevice.stopEmulatorOutput, since the
getEmulatorOutput is in the ITestDevice.
In TestInvocation, add verify of device.getIDevice() != null for
defense.

Change-Id: I70a2edbc5acd42e1d616b5b35b553704423894ca