Don't cache root file entry in TestDevice
This was causing syncFiles() to fail on MNC due to the implementation of
ddmlib.  It seemed that the mapping of files was getting corrupted due
to files being deleted, and this was causing the method to look for
the dir under /sdcard/sdcard/, not /sdcard/.  If we don't cache the root
file entry, this problem goes away.  It does not look like this is a
heavily used function so risk seems low.

Bug: 20823584
Change-Id: I546f89bbc92c1d9433527d280bc0c86e557bdb80