Create target directory in FileUtil::recursive{Copy, Hardlink}.
Create the destination directory in recursiveHardlink and recursiveCopy if it
does not exist. This makes behavior consistent with recursiveDelete.

This CL fixes broken FolderBuildInfo::linkOrCopy which assumed that
recusiveDelete followed by recursiveCopy would restore directory contents.
Since recursiveDelete deletes the target dir and *Copy does not create the
destination dir if it doesn't exist, the *Copy always fails.

Bug: 18247399
Change-Id: Ib88e47ea2258ffad30a12f9f747b862a01fbb15c