CommandLine objects now contain line number and file.
CommandFileParser was changed to keep track of the line number as it parses
through a command file. This line number and the file are saved inside of the
CommandLine object containing the expanded command line. The stored line
number and file always corresponds to the location of the original command
line before any expansion.

The purpose of this change is to provide better support for configuration
validation tools that make use of the CommandFileParser class. This change
should also make it possible for TradeFed itself to provide more verbose
and helpful error messages.

Change-Id: Id2b73a1683710e005b495964a6418f0134cc726b