fix free space detection by relaxing regex for df output
When we detect free space, we cal df with a parameter indicating the
mount point of external storage. This parameter comes from
environment variable EXTERNAL_STORAGE set on device, and may be a
symlink path. The df output may be using the real path of the mount
point, which could be different from what's originally provided. This
change accomodates for such difference.

Bug: 25260612
Change-Id: I681ab8018a207d08563643a934e22ebe8df3eff7