avoid getting stuck on process that won't terminate
When external process is stuck in uninterruptible sleep, it's not
possible to terminate the process, and associated tracking threads
spawned by RunUtil. This change tries its best to decouple the
invocation thread from hung process, at the expense of leaving some
stray threads around, but allowing the invocation to fail.

Bug: 26314635
Change-Id: I9d2edcc8aae58f1ab502459469a6831b904c3844