Adds support to wildcards and log messages.
Some devices' shell returns an extra white character at the end of the
filenames when listed using 'ls'. Added trim() to alleviate this issue.

Surrounding the argument of ls with single quotes makes it literal,
therefore wildcards don't work. Removed the quotes to allow to better
filter the content of a single directory. Besides, patterns with
wildcards are widely used and common.

Change-Id: I58d3f5bf7e33befd2569df278149b26a6c0a3f4f