Adding a new preparer to initialize attenuator
When a test running with RF attenuator is hit by exception,
host system upgrade, Tradefed Upgrade, power otage etc.
The attenuator setting will be in unknown state.
This preparer can make sure attenuator value is reset before the
test start, epsecially before the pairing between phone and watch
started. By defailtm this preparer is disable, user need to enable
it in command file and pass in the ipaddess list for all the
attenuators in the physical setup.

Change-Id: Id0d4a6bd75c1d0ea885fc14d05fab2aae51a180b