Migrate ColossusLogSaver to AndroidBuildAPILogSaver.
Previously the unit-runner uses ColossusLogSaver and sets the stage path
to nfs (/auto/android-test/www/test_logs), and it's not deleted after
testing (this may be a leaking, check /auto/android-test/www/test_logs/tradefed)
We are migrate ColossusLogSaver to AndroidBuildAPILogSaver. And we use
/tmp/stage-android-build-api as stage root folder. The folder itself can
not be deleted, but the actual invocation folder in it will be cleared.

Bug: 31152039
Change-Id: I6afc6cd4de80b224ed0c7d0e268c3896a1b832e6