Avoid deadlock in FileDownloadCache
Trying to synchronize access to File objects can result in deadlock when
trying to remove a file from the cache that is being downloaded at a time,
resulting in a thread holding onto the mCacheMapLock, and blocking other
threads from downloading files. This can happen frequently when:
  * A large number of invocations begin downloading files around the
    same time.
  * Not enough space in the cache. If a file is downloaded twice, it
    will increment mCurrentCacheSize twice, but will only be able to
    decrement it once. This change doesn't completely eliminate that
    issue, so it may be worth fixing that in another change.

This change adds an abstraction of file locks that are independent of
mFileCacheMap, and avoids deadlock by not removing files that are
currently being used by another thread.

Test: Added functional test that fails previous implementation
Bug: 35861062
Change-Id: I105b569d1c6158c8cf4fc261e2d1183395181610