TfSuiteRunner: Add in option suite-config-prefix.
Currently TfSuiteRunner will only look for configs that are prefixed
with 'suite' for a suite tag.  I'm reversing the logic so that by
default we look for all configs and if a user wants to, they can pass in
a prefix to limit the search space for the configs with the suite tag.

BUG:36515064
Test: Tested on basic suite that suite runs and finds all the configs
and when using the prefix also looks at configs I expect.

tf >run google/template/lab-base --template:map
test=google/suite/bvt_suite --branch git_master --build-id 3857424
--build-flavor angler-userdebug

-and-

tf >run google/template/lab-base --template:map
test=google/suite/bvt_suite --suite-config-prefix google/test/androidbvt
--branch git_master --build-id 3857424 --build-flavor angler-userdebug
Change-Id: I682f6ffcb8ab8b00dd5c25b6b4475d30997feab4