Move output-test-log directly to xml config
To avoid launcher to inject option that does not exist
for the version of the child process. We move the option
to the xml config directly to be always in sync with the
branch.

Test: local run
Bug: 37706926
Change-Id: I54659ce6efe5b02628cd337703ddf88233946df4
(cherry picked from commit 80a5dae53ac6465db5cb8a84ddf0d190b3e68b5b)