switch to use ro.product.board for "Product"
What TradeFed is calling "Product" is really "board", which is
a misnomer from the beginning. On device side, the meaning of
various properties are poorly documented to begin with.

The intention of this field is to track a device type across
adb and fastboot. It originates from the enforcement by fastboot
that "requires board=" field in android-info.txt in the image
zip file must be identifical to what "getvar product" returns by
device bootloader: TradeFed needs to be able to track the same
value when device is in adb mode, and "ro.product.board" clearly
fits that purpose semantically

Bug: 64083540
Test: existing unit tests
   validated flashing on marlin, sailfish, angler, bullhead
   checked property field on taimen, walleye

Change-Id: I51b1d12eaa6dc56867eaac82d07944bd3396200e