Update get result server args to catch all Exceptions.
I just ran into an ssl.SSLError exception where the read operation timed
out. I initially added it to the exceptions we catch but realized we
just want to make sure we can urlopen no matter what and that a general
exception catch is better (in terms of maintenance so we don't keep
adding new exceptions) and not risky since we're only trying on the
urlopen call.

Bug: 69316075
Test: upload to sponge still works
Change-Id: I26d7e4b08b5a9583bd35091fc18f01e16fd3dab8