atest: Add --rebuild-module-info flag
Atest caches the module-info build artifact between runs.
This results in staleness that can occur after a repo sync or
when a new test is added.

This change adds a new flag to atest to force the re-initialization
of the module-info.json file. If a possible error occurs, we prompt
the user to attempt running atest again with the flag in order
to resolve the issue.

Bug: 71510175
Test: Changed HelloWorldTests to HelloWorldTests2.
      ran `atest HelloWorldTests2`, the error occured.
      Copy pasted the suggested command line and the test passed.

Change-Id: I2e142d2df4bb70cd1a928378c02374114f6e4b29