Add atrace metric collector.
Upload the trace files collected during the test in sponge.

This is needed inorder to switch the Jank test to use
AndroidJUnitTest because there is separate jank test run with
trace collection during the test and uploading it to sponge.

Also if the script path exist process the trace files for
more detailed metrics from the test:

Bug: b/70337662

Test: OK (4 tests)

Trace files collected during the test are uploaded in the
sponge link as expected and scripts from the path are
executed against the trace files and metrics are collected
as expected.

Change-Id: I732a8a28de33ae26ad474ced2325f7995f03ae81