Fix module:class lookup where class outside module.
If given a module:class reference, we would only look for the class
under the module config dir. But some classes exist outside the module
config path, so now we re-search from repo root when we don't find. It's still
best to look first under the module path for speed purposes though.

Test: ./cli_translator_unittest.py
      atest CtsInlineMockingTestCases:com.android.dx.mockito.inline.tests.MockFinal
Bug: N/A

Change-Id: I1adb95895dfcf1153f485674ccffaaede4e4e35f