atest: Comment out unsupported test references.
After the refactor, atest will print out to the screen test references
we do not support and can cause more confusion than educate. Let's drop
that messaging and only add in test reference types we support and
comment out those that we plan to support but currently do not.

Bug: 73559221
Test: atest CtsAppTestCases:android.app.cts.AlarmManagerTest does not
print warning anymore.

Change-Id: I724c3c7f5fd023038bdf201df584fe6c6b7ecd78