Introduce a FileProperty for BuildInfo
Fileproperty will allow us to handle some file a bit
differently depending on the property. This allows us to
customize how we handle tests dir coming from the
build_provider, or the ANDROID_OUT_* variables.

This is also intended to help with sharding file sharing
optimization later on.

Test: unit tests
run cts-dev -m CtsGesture (with and without lunch)
./tradefed.sh run cts-dev -m CtsGesture --use-sandbox
Bug: 73343867

Change-Id: If1713648b868374d207751fb9ba201975ebab462