Have AtraceMetricCollector work with postprocsesing tools.
After collecting the trace, allow the metric to use self-contained
binaries to process the traces. A fair amount of existing trace tooling
is not in Java, and this allows reuse of the tooling without porting.
The subprocess results are managed in the same way that the
PythonUnitTestRunner works, it will launch the process and then extract
metrics and/or result files from the output text according to the
defined grammar.

Bug: 73541075
Test: 4 new unit tests. Run:
      tradefed.sh run singleCommand host -n --class
      com.android.tradefed.device.metric.AtraceCollectorTest
Change-Id: I745288edd7e0fc4ca1cd9590ce50a54f107f9933