Renaming atest configs for longevity.
Renamed the atest configs to test longevity to reflect the local nature
of the configs. This will serve two purpose, first, the name makes it
clear that the configs are not to be used in atp continuous testing and
second, it removes the possibility of collision of config names in
tf-prod-tests and google-tf-prod-tests which may create confusion.

Test: atest longevity/test -- --template:map metrics_collector=google/longevity/metrics-collectors --template:map preparers=longevity/preparers
Change-Id: I370646a91943f94d50b7edfe4e513a9ad1993109
Bug: 76212634