Support module-arg in TEST_MAPPING files
This change allows TEST_MAPPING json file to include module-arg for each
test module. The module args will be added to suite runner for each
test module.
An example of a test with option in TEST_MAPPING looks like this:
{
  "name": "test2",
  "options": [
    {
      "instrumentation-arg": "annotation=android.platform.test.annotations.Presubmit"
    }
  ]
}
When TradeFed runs a suite for this TEST_MAPPING file, the command line
will be like:
tradefed.sh run ... --include-filter test2 --module-arg \
  test2:instrumentation-arg:annotation=android.platform.test.annotations.Presubmit

Bug: 79529839
Test: unittest
Change-Id: I58aaf339f810a5666277abb1aa2703cf6f00d97d