Refactor LogcatUpdaterEventParser into LogcatEventParser
LogcatEventParser is generalized to work in other use cases.
Events regarding system update are no longer hard coded into the constructor.
Logcat output is immediately processed into events and put into a queue.
This omits the unnecessary file backed buffer that was previously used.
The waitForEventAsync function is also removed, as nobody was using it.

Bug: 80206402
Test: tools/tradefederation/core/tests/run_tradefed_tests.sh --class com.android.tradefed.util.LogcatEventParserTest
Change-Id: I321b3e3d4620527b5d9fb3abcc8e910bd61ff79b