Modify the NoisyDryRun behavior for sandbox
When a command is sandboxed (--use-sandbox), only load the
non-versioned part of it, as it is the only basic thing
that we can validate at that point. This ensure that the
command can at least load against the current TF deployed.

The rest for now is up to the command itself and not to
the lab setup so it can be more easily fixed when it fails.
(without requiring a new deploy).

Test: unit tests
Bug: 75033502
Bug: 110545254
Change-Id: Ia98cfe93cdb07abd80ade68e383a4863ea80b275