Suite Sharding Memory improvement
Drop the filters once not useful anymore during sharding.

This will save several GBs of heap memory because of all
the filters we carry (especially in retry). This also
improve the init speed of sharding by 10 seconds.

Test: unit tests
Retry a full run with a bunch of failures:
run retry --retry 0 --shard-count 6
Bug: 77648991

Change-Id: I5df4580b80aa71df5afec926667303125cf7b716