Handle sharding suites for Tcp device
Sharding of tests occurs before tcp device connection, so
for a ITestSuite that tries to get the abi information we
need to get the device setup first.

Change the flow of device setup during sharding, to ensure
we have a chance to see the device.

Ensure we don't call tearDown in parent sandbox too because
setup wasn't called, it might results in odd behavior.

Test: unit tests
run the failing config from the bug
https://sponge.corp.google.com/invocation?tab=Test+Cases&show=FAILED&show=INTERNAL_ERROR&id=9484139e-8f43-4805-a78f-91e61fef4ca6
Bug: 70778109
Bug: 112385225

Change-Id: Ibb24305742ba5f0eae8a1995fdf517566ed27043