atest: fix osx socket bug
OSX handles non-blocking sockets in a different manner than linux/windows.
So we need to block (no timeout) on the recv() function. This is fine, because
we've connected at this point, so we'll either get data or if tradefed
exits for any reason we'll unblock as the connection is closed.

Bug: b/112382535
Test: ATEST_PRETTY_RESULT=1 atest CtsActivityManagerDeviceTestCases:ActivityManagerMultiDisplayTests
Change-Id: Idd6b69549558a909394c779b2574140c40634731