Fix pushing a file into a directory
PushFilePreparer contract allows to push to an existing
directory a file. In such case we push with the same name.

Also ensure that not two files get pushed to the same
remote location as they would override each other.

Test: unit tests
atest -v perfprofd_test (with couple of variation in AndroidTest.xml)
Bug: 113512417

Change-Id: I8a854a65684a826ddd106c0cea63059849e5dde7