atest: TEST_MAPPING warn if specified tests is not a module or in the right suites.
TEST_MAPPING local runs will work if the user specifies tests via
class or packages but they will not work in TreeHugger. This is
not obvious to the user so warn them in this situation.

Also if the test module does not belong to device-tests or general-tests
the tests won't work in TreeHugger, so warn for this situation
as well.

Bug: 117469567
Test: local atest runs.
Change-Id: Ib7f27559bff847f399a7b753db3d7985dc45ab55