Adjust test option merging logic for tests in TEST_MAPPING file.
include-annotation applies to all tests, which is use to only include tests
with certain (Presubmit) annotations. That can cause tests don't rely on
the annotation to be skipped after the merge.
exclude-annotation can be used to skip flaky tests. Some test packages don't
use the annotation so there is no need to use the exclude-annotation option.
By keeping any exclude-annotation option, we try the best to not introduce
flaky tests unexpectedly.

This is a temp fix until a suite run redesign to support suite to run tests
with different options to run in its own test run.

Bug: 117880789
Test: unittest
Change-Id: I55bedf0ea443777505b4e5b73d061152e1f7a0c9