Allow passing run name to JUnitXmlParser
JUnitXmlParser, which is currently used by GradleTestlogForwarder,
gets the test run name from the "name" attribute of the <testsuite>
tag. It appears that XML results from Gradle contain a single test
suite, whose name is the same as the file name (with the "TEST-" prefix
and ".xml" suffix removed).

GTestXmlResultParser uses a name specified in the constructor.
GTestTestlogForwarder passes it the XML file name with the extension
removed.

As an initial step towards unifying the two parsers, this change adds a
constructor to JUnitXmlParser which enables the caller to specify the
run name.

Change-Id: I137cddafdd4d5c674e786bceae73483ddc2d9f83
Issue: 120047628