Remove RunUtil.RunnableResult closeStreamAfterRun argument
closeStreamAfterRun's behavior is currently slightly wrong - in that if
either stdout or stderr streams are provided but not both, the other
stream that RunnableResult creates is never closed. However, it's
unnecessary in any case; it's better to just track in RunnableResult
whether it created the stream(s).

Bug: None
Test: m -j tradefed-all &&
tools/tradefederation/core/tests/run_tradefed_aosp_presubmit.sh

Change-Id: Ic48d9d17af4a6abfe56d9fb11152b7c26cee656c