Support SKIPPED GTests
I encountered an error when a ran a test using 'atest' which happened to
use the new GTEST SKIP() to not run.

https://github.com/abseil/googletest/pull/1544

It turned out that GTestResultParser did not handle the SKIPPED message
that is emitted, and it later raised an exception when it received a
message it did understand, but it thought was out of sequence.

This patch alters the parser so that the SKIPPED message is handled, and
the test is reported as ignored.

I also added a new unit test to ensure that SKIPPED tests are handled
without an error in the parser.

Test: run_tradefed_tests.sh
Bug: None
Change-Id: If8a9c8c0256293217106b302c224b0b7ccf0b353