Adding target preparer for switching to primary or system user.
Some modules are testing features that are only applicable for user 0.
UserTargetPreparer encapsulates logic for switching to system (user 0)
before module execution, and switching back to user we were in before the
execution.

For most intents and purposes, primary = system user, but since these
two flags are separate in the framework, it's better to make intentions explicit
and provide support for both.

This enables us to run the suite in a secondary user. Teardown switches
back to a secondary, so that other modules are not affected.

Change-Id: Ifd53d51302f363240a3016819877eb9fccdb56d6
Fixes: 120287031
Test: SwitchUserTargetPreparerTest in tf unit tests
Merged-In: If2e1f558e67d5e83fa68933a0ef1367583bdcb5f