ATest: Support user customize OUT_DIR_COMMON_BASE
When using OUT_DIR_COMMON_BASE to define a customized out directory,
ATest will fail to build module-info.json since it passed a reletive
path to the build system. This patch is to fix the issue that ATest
can always build successfully no matter OUT_DIR_COMMON_BASE is set
or not.

Bug:123555460
Test: m atest
      atest hello_world_test
      export OUT_DIR_COMMON_BASE=/usr/loca/tmp/out/
      atest atest_unittests

Change-Id: I8b41fcad36f73095bdf4f7df0f4128efb76f867c