Do not run presubmit tests in postsubmit group
We used to run all tests in both presubmit and postsubmit group when
the test-mapping-test-group is set to postsubmit. This offers maximum
coverage. However, since lab has already scheduled postsubmit runs for
the presubmit test group, there isn't a need to run the presubmit test
again in the postsubmit group.

This change remove the logic to include presubmit tests in postsubmit
test group test.

Bug: None
Test: unittest
Change-Id: Ia3085a9a9554c13d2bac27d0c81b8a94bed7069c