ATest: Running integration test with options.
Previously, integration test takes more than half an hour to complete all tests.
Instead of running the whole process of find, build, install and test, we can
leverage the "--dry-run" option to focus on validating the process of target
finding as well as the command sent to TF, so that more time can be saved and
increase the test coverages.

Bug: 128876719
Test: tools/tradefederation/core/atest/atest_integration_tests.py --dry-run
Change-Id: I1da094ecdb5c9394939b4b3240551f6843956f03