ATest: Ignore test status not in the status list for result report.
Atest cares only 3 status: passed, failed and ignored. Previously when
a test exits with ERROR(not failed), the result_reporter will throw
RUNNER ERROR which is not a key in _SUMMARY_MAP_TEMPLATE, and results
in run time error. This CL is to ignore those not in our desired status
gracefully.

Bug: 128583603
Test: m atest && atest-dev hello_world_test HelloWorldTest
      cd frameworks/base/startop/iorap && atest-dev

Change-Id: If03306c9ce81e2b761da22f6eeaa893a49fd9e26