Atest: Create a cache file for locate.
"updatedb" is not available on Darwin platform. Without it, Atest will
not be able to utilise locate command to speed up finding files.

This CL provides a rewritten "updatedb" script for Darwin to create a
database readable for "locate" command, which can read the database
without traversing the filesystem over and over again.

The purpose of it is to copy it to HOST_OUT and rename it "updatedb" in
Darwin, and invoke tools/atest_updatedb.py in both Linux and Darwin.

Bug: 132217764

Test: $ source build/envsetup.sh; lunch
      $ cd tools/tradefederation/core/atest/tools
      $ ./updatedb_darwin.sh -h # helping message
      # generate a database in $ANDROID_HOST_OUT/locate.database
      $ ./atest_updatedb.py

      # speed comparison:
      $ time find $ANDROID_BUILD_TOP -name gtest.cc
         (0m7.880s)
      $ time locate -d $ANDROID_HOST_OUT/locate.database gtest.cc
         (0m0.645s)

Change-Id: I0afbc23448d37d90cc8ac128db60fe179fe70f6c