ATest: store test command result for verification.
This CL enables an option "--update-cmd-mapping" that implicitly enables
dry-run to store test runner command which can be updated when next time
the test runners return a different command. When encountering contradictive
command result of a test, Atest will interactively ask the users to update
the result or skip updating it. This implementation ensures the correctness
of test runners and uncovers potential issues before  releasing prebuilt
binaries.

Bug: 133475911
Test: atest-dev hello_world_test --update-cmd-mapping ;
      check content of
      'tools/tradefederation/core/atest/test_cmd_result.json'
      python tools/tradefederation/core/atest/atest_integration_tests.py \
      --update-cmd-mapping ;
      check test_cmd_result.json
Change-Id: Icdf03e6334a2c99a334621b3a668c16261b2c285