Improve logcat-on-failure accuracy
Reduce the default number of buffered logcat line printed
when we start our collection. This avoids the full buffered
logcat to have to print before we can get a log.
This will result in much smaller logcat-on-failure as well
as not missing later stack when the printing of buffered
logcat is slow.

Test: unit tests
local tests
Bug: 128425333

Change-Id: Ie1bd06784d5d09479099a0989316e31ac904baf2