Remove tradefed's dependencies on the java.xml.bind module.
This dependency (for the Java 9 module system) was added in
http://r.android.com/524977 citing an import in StreamUtil.java. That
import was removed in http://r.android.com/649822, and there are no
longer any references to any of this module's packages in tradefed.

As well as removing an unused dependency, this is good because the
module is deprecated, and is removed in OpenJDK 11.

Test: EXPERIMENTAL_JAVA_LANGUAGE_LEVEL_9=true make tradefed tradefed-tests
Bug: 131683177
Change-Id: Ib39d2a4ce510689ed7d90248b290e475c7cc4f5f