Add onTestEnd with test description
To add log files with names that references the test name, it's
necessary to add a TestDirection argument to the onTestEnd function. By
default, the original definition of onTestEnd will be used. If the
version of onTestEnd with a TestDescription is overridden, the other
overloaded definition will be ignored.

Bug: 134790621
Test: tradefed.sh run commandAndExit host \
--class com.android.tradefed.device.metric.BaseDeviceMetricCollectorTest \
--method testOnTestEndWithTestDescription -l verbose
Change-Id: Ia748d5c073fcd00b2f29821c2d722d5b4c56ee9a
Signed-off-by: Maggie White <maggiewhite@google.com>
Merged-In: Ia748d5c073fcd00b2f29821c2d722d5b4c56ee9a