Allow empty values to be specified for options.
Add a limit of 2 to the split statement that tries to identify key-value
pairs. This allows a trailing empty part to be captured after the :=
key-value separator, which thus allows a key to be given an empty value.

Also improve testing of the handling of empty strings, both for simple
values, for lists, and for maps with key-value pairs.

Tested using

mmma -j$(nproc) tools/tradefederation/core/ && \
tools/tradefederation/core/tests/run_tradefed_tests.sh \
--class com.android.tradefed.testtype.suite.SuiteModuleLoaderTest

Test: Included.
Bug: None.
Change-Id: I6b0356bfd0734ba40171a3572f82199893cf4514