Atest: tolerate finding targets in an inexistent dir.
Previously if the module_info.json is outdated, finding a test target
will result in searching in an inexistent directory and raising
subprocess.CalledProcessError. This implementation is to check the
existence prior to searching in it, and eventually the users will be
guided to run "--rebuild-module-info" to resolve the issue.

Bug: 138242352

Test: source build/envsetup.sh && lunch
$ mv test/suite_harness/tools/cts-tradefed test/suite_harness/tools/cts-tradefeD
$ atest-src PwrStatsUtilHelperTest
$ mv test/suite_harness/tools/cts-tradefeD test/suite_harness/tools/cts-tradefed

Change-Id: Ia05727406148f5412209fdc49a72f91d4ad464a1