Allow preparers to install apk from remote zip.
This change allows following preparers and utilities SuiteApkInstaller
to try to stage a file from remote zip dynamically.
PushFilePreparer, SuiteApkInstaller, BuildTestsZipUtils

Due to build limitations, some test APK files are not under the test
module directory. That leads to test failure when partial download is
enabled.
With this change, these tests can work for partial download (when
test-config-only-zip is set).

Bug: 73786521
Test: local run, unittest
Change-Id: I65e75fa9e90d69f1b27db3604702cd202dbbc40f