use java binary that launched current process for subprocess
As a first step of consolidating use of java binary in TF, we
are moving away from using the java binary from path; instead
we are using the same java binary that launched current process
for all subprocess test execution needs.

In the future, some of these use cases will be changed to be
using java binary based on the API level where subprocess
artifacts are obtained (or they'll come pre-packaged with java
binaries).

Bug: 122862400
Test: existing unit tests
Change-Id: I9dc6907cba6af65335ad7e490bb67fbc5b578657