ATest: Replace static make command with a get method.
The make command was a static variable which contains the relative path
of soong_ui.bash, as a consequence it could generate a wrong command when
importing module atest_module_info and changing working directory. This CL
fixes the issue by refactoring the variable as a get method, so that no
matter what the cwd is, the make command can always be correct.

Bug: 139264620
Bug: 136623722
Test: m atest ; atest-dev hello_world_test
      m atest ; atest-dev RunSettingsRoboTests
      m aidegen ; cd tools/tradefederation/core ; aidegen-dev -a

Change-Id: I46959aa3325e321f09b29f16e551c5c5fee80144