CommandScheduler: Tweak warning message.
The commands that it was warning about can be extremely long.
This resulted in messages of the form:

The command [...] will be rescheduled.

where ... was more than a screen full of text.

I've moved the command to the end of the sentence so that a
human can at least understand what the message is telling them
without having to hunt for the end among a wall of text.

Test: Treehugger

Change-Id: I6a4d8c9d884f04e23993bbc54d1f1e49f63c217b