This adds a check to handle GBenchmark errors.
Previously, TF did not parse GBenchmark results to check for errors, but
this becomes problematic when failures are reported through those
results.  Thus, this adds a simple check for the existence and value of
the "error_occurred" key, and if that is set to true, then it will cause
a failure on the test and attempt to extract the error message.

Bug: b/134576138
Test: Unit test
Change-Id: I723ca6277e03a3bcd3971a0a022baebc066bd06a