Allow BaseTestSuite to load the given TF configs.
In the current design, ATest has found the config in the android source
tree but still depends on TF to do finding the configs, which causes a
lot of overhead finding the configs.
The CL is to pass the found config path directly to TF to reduce the
overhead.

Bug: 141090547
Bug: 131429500
Test: unittests.
Change-Id: I16221518304f984235819b06822935422e5cf6dc
Merged-In: I16221518304f984235819b06822935422e5cf6dc