Support known network device by IP
In some cases, devices exists on reachable network and we
want to use them instead of a local device. Allow to
support this.

Have a pool of known ip for gce and tcp device in order to
match the current split.

Test: local runs, unit tests
Bug: 143531490
Change-Id: I48bccb21a9ee81882969f063f80592ead331631a
Merged-In: I48bccb21a9ee81882969f063f80592ead331631a