Adds a variety of refinements to isolation runner.
When initially trying to run some of the robolectric tests under this
new runner, there arose some unforeseen issues (like dependencies
including things in the java.lang package) that for now are mitigated by
the runner.  Additionally, the change to using RunUtil inadvertantly
disabled output streaming from the subprocess, which this CL fixes.
Lastly for the major changes, this changes the qualification of what
counts as "the correct annotations" which excludes classes which specify
the "RunWith" annotation, but use a runner that does not subclass the
JUnit4 runner.

Edit #1: I moved the fakeExecution out of HostUtils into a proper class
that extends the JUnit Runner API.  This makes the flow when dry running
a test quite a bit more similar to the flow when running for real.

Bug: b/130198687
Test: incomplete, but no regressions in existing
Change-Id: I4a6fda942e47978506a6ddc82c5f8a03dc04b690