Improves HostUtils and migrates HostTest to use it
This enables HostTest to take advantage of the parallelized JAR
searching enabled in HostUtils.  It took a few modifications to
HostUtils, however, to line the behavior up with the old implementation.
Most notably, our unit tests currently enforce an implementation detail
that the class list returned is in order by which they are found in the
archive when this is not a part of the function contract.

Edit #1: Changed to allow anything that runs with anything that extends
`org.junit.runner.Runner` to be considered a test instead of the more
restrictive subclass check before.

Bug: b/147308961
Test: unit tests still pass (except the thread count one)
Change-Id: If3c55f5789c998ceb8bd0d7ac2a2a3cd9b99a815