Add backup and restore functionality
Need a way to save original system files when pushing a mock system
file on top. The mock will call the original for certain calls, so it
needs to be available on the device still while the test is running.
Adding a backup command option that
allows the system and vendor partitions to be remounted as writable, and
allows the commands to run with needed permissions to edit the system
files both to a backup location and back.

Test: tradefed.sh run google/template/asit --template:map reporters=google/template/reporters/blackbox --log-level verbose --log-level-display verbose --template:map test=google/test/framework/media/widevine/tests/widevine-modcrypto-unit --wifi-network GoogleGuest --use-build-api --no-use-sso-client --branch git_rvc-release --build-flavor bonito-userdebug -s 93GAX07RDD --com.android.tradefed.targetprep.RunCommandTargetPreparer:run-command "sm set-isolated-storage off"

Bug: 139085609
Change-Id: I81132bd8a9cc9a32aea546b7c977b17f7c58f643