ATest: Add vts-core-tradefed to test dependency.
If test module belong to vts-core suite, adding vts-core-tradefed to
test dependency due to it might using vts-core's test harness, in oder
to prevent class not found error, adding more build targets.

If a testable module belongs to vts-core suite, it may use vts-core
test-harness and result in class-not-found error. This CL added
vts-core-tradedfed to dependency list to prevent from happening the
error in the test running time.

Bug: 149838888
Test: ./tools/tradefederation/core/atest/run_atest_unittests.sh
      apply aosp/1236886 ; m atest ; atest-dev KernelSelinuxFileApiTest

Change-Id: Ib45ef7bb0ba40b3b42ac4ba4e2f72c2dc8c446f3