Retry when `Permission denied` is returned (2/n)
My test shows that with one more retry, the external storage will be
mounted correctly in most cases even when 'Permission denied' is
returned previously. See b/150749023#comment3 for more details.

This will reduce the chance of unnecessary recovery/reboot due to
failed #waitForStoreMount and greatly improve the steadiness of host
side tests.

Note retryOnPermissionDenied is set to 1 so we won't retry endlessly to
impact the boot tests.

testWaitForStoreMount_PermDenied is also modified to check if
'retryOnPermissionDenied' works correctly.

Bug: 150749023
Test: atest StagedRollbackTest#testBadApkOnly
Test: tradefed unit tests

Change-Id: I651ad8021b880f269f5e2d596a8d79679429164a