Add use-zip64-in-partial-download in HostOptions.
Partial download isn't enabled in qt-dev branch. Once we add host
options in the host-config.xml, it will break becuase this branch can't
know the option(use-zip64-in-partial-download). Also, it has code conflicts
while cherrying pick it to qt-dev. This change is to solve this problem
so that it will not break once we add host options in the
host-config.xml.

Bug: 150588118
Test: unittests.
Change-Id: I5f54433548b32a5f1a3be5d15cb7792a40330964
Merged-In: I27f8148b9ee31fb3b4bf0aa8c79b90df02b36c44