improve the wait logic in nested ssh bugreport
If this is the first time running an adb command on remote VM,
the nested virtual device might not become visible to the adb
server immediately. This is due to the implementation of
adb_connector on remote VM.

This patch prefixed the bugreport command with a
`wait-for-device` to ensure that the device is visible first
or until the preset 15m timeout. The previous retry loop is
also removed now that the cause of the behavior is clear.

Bug: 154613608
Test: updated unit test
      http://sponge/4503f5cc-0fc1-415e-9cb0-06e06317ae53
Change-Id: I2269be1085be23ecf0e97658502d387cbcc0d9f1