Add support for installing APKs from directories
Add support for specifying directories that contain APK files. This
enables users to install a set of APKs from an external directory which,
in turn, can be resolved using TF's remote file resolution mechanism.

This addresses the case where the exact set of APKs is not statically
known and avoids having the user write multiple test configs for each
combination. For example, an app can have different APK splits that
depends on the target device it's being installed on.

Test: m tradefed tradefed-tests && ./tests/run_tradefed_tests.sh
Bug: 156254548
Change-Id: I7e2195ce10b42081039331fd4dfe7f7d31eddf49